Trying to organise
Organinsing my pc and desktop and access to github to make it easier to work through and with.  Also had far too much linked into github, took so long to look through, loadetc so am trying to optimse everything. Please bare with me, it takes me a while to learn new stuff. Why I am looking to learn how to use android studio is not in read me files in a folder dedicated to all that is android studio.  It is open to the public so anyone can grab anything they need, although it may be a while before I contribute anything that maybe usefull :) I will get there, really want to build this platform.  See readme file in .andyandroidstudioprojects